# docs/swagger.yaml
...
openapi: 3.1.0
paths:
...
  /manage/badge-categories:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BadgeCategory"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Badge Category Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Badge Category Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/badge-categories/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBadgeCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeCategory"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Badge Category Resource
      security:
      - BasicAuth:
        - ADMIN
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeCategory"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Badge Category Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/centres:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Centre"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Centre Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCentreRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Centre Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/centres/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCentreRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Centre"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Centre Resource
      security:
      - BasicAuth:
        - ADMIN
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Centre"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Centre Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/challenges:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChallengeDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Challenge Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Challenge Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/challenges/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChallengeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Challenge Resource
      security:
      - BasicAuth:
        - ADMIN
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Challenge Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/games:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Game Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Game Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/games/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Game Resource
      security:
      - BasicAuth:
        - ADMIN
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Game Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/level-definitions:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LevelDefinition"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Level Definition Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLevelDefinitionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Level Definition Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/level-definitions/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLevelDefinitionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelDefinition"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Level Definition Resource
      security:
      - BasicAuth:
        - ADMIN
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelDefinition"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Level Definition Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/metadata:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Metadata"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Metadata Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMetadataRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Metadata Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/metadata/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMetadataRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Metadata"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Metadata Resource
      security:
      - BasicAuth:
        - ADMIN
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Metadata"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get
      tags:
      - Metadata Resource
      security:
      - BasicAuth:
        - ADMIN
  /manage/modules:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManageModuleDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Module Resource
      security:
      - BasicAuth:
        - ADMIN
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Module Resource
      security:
      - BasicAuth:
        - ADMIN
...
  /manage/submissions:
    get:
      parameters:
      - name: status
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: List
      tags:
      - Challenge Submission Resource
      security:
      - BasicAuth:
        - ADMIN
...
components:
  schemas:
    BadgeCategory:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        displayName:
          type: string
        description:
          type: string
...
    Centre:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        active:
          type: boolean
    ChallengeDto:
      type: object
      properties:
        challengeId:
          $ref: "#/components/schemas/UUID"
        moduleId:
          $ref: "#/components/schemas/UUID"
        badgeCategoryId:
          $ref: "#/components/schemas/UUID"
        displayName:
          type: string
        description:
          type: string
        xpValue:
          type: integer
          format: int32
    CreateBadgeCategoryRequest:
      type: object
      properties:
        displayName:
          type: string
        description:
          type: string
    CreateCentreRequest:
      type: object
      properties:
        name:
          type: string
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    UpdateBadgeCategoryRequest:
      type: object
      properties:
        displayName:
          type: string
        description:
          type: string
    UpdateCentreRequest:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
    UpdateChallengeRequest:
      type: object
      properties:
        metadataId:
          $ref: "#/components/schemas/UUID"
        displayName:
          type: string
        description:
          type: string
        badgeCategoryId:
          $ref: "#/components/schemas/UUID"
        xpValue:
          type: integer
          format: int32
...
  securitySchemes:
    BasicAuth:
      description: Authentication
      type: http
      scheme: basic
info:
  title: Wishaw YMCA Esports API
  version: 1.0.0
  description: Backend API for the Wishaw YMCA Esports Login App
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value